+++ tester.hpp
@@ -6,7 +6,7 @@ namespace tester {
 
 struct Config {
         void (*const sort)(std::vector<int> &);
-        const int number_of_tests;
+        const size_t number_of_tests;
         const size_t length;
         const int min;
         const int max;
@@ -14,7 +14,7 @@ struct Config {
 
     public:
         Config(void (*const sort)(std::vector<int> &),
-               const int number_of_tests, const size_t length, const int min,
+               const size_t number_of_tests, const size_t length, const int min,
                const int max, const bool critical)
             : sort(sort),
               number_of_tests(number_of_tests),
@@ -23,7 +23,7 @@ struct Config {
               max(max),
               critical(critical) {}
         Config(void (*const sort)(std::vector<int> &),
-               const int number_of_tests, const size_t length, const int min,
+               const size_t number_of_tests, const size_t length, const int min,
                const int max)
             : Config(sort, number_of_tests, length, min, max, false) {}
 };
